requirements.txt
Flask==0.12.2
Pillow==4.1.1
